requirements.txt
scipy
tqdm
numpy
joblib
pandas
ipython
librosa
scikit-learn
matplotlib
argparse
kaldiio
PyYAML
pickle4
wandb
git+https://github.com/Legisign/Praat-textgrids
jupyter
seaborn
flask